+++ src/Slide/index.js
@@ -1,9 +1,1 @@
-/**
- * Slide Component Style for SaltUI
- * @author changming
- *
- * Copyright 2014-2017, SaltUI Team, Alinw.
- * All rights reserved.
- */
-
-export default from './Slide';
+export default from '../Carousel';